feat: decouple kyc pages (#388)
* Update `me` response

* Fix content-type check

* Fix Rapyd API URL

* Update account model table name

* Update default DB URL in zod schema

* Update `isAuth` guard

* Check for session in `me` controller

* Update Rapyd service and controller

* Fix route paths

* Update frontend middleware

* Update countries and documents fetching

* Update user response type

* Decouple KYC pages

* Remove console logs

* Refactor countries and documents API calls
