+++ packages/wallet-backend-update/src/user/controller.ts
@@ -3,6 +3,7 @@ import type { NextFunction, Request } from 'express'
 import type { Logger } from 'winston'
 import type { UserService } from './service'
 import type { SessionService } from '@/session/service'
+import type { User } from './model'
 
 interface UserFlags {
   needsWallet: boolean
@@ -23,10 +24,17 @@ export class UserController implements IUserController {
 
   me = async (
     req: Request,
-    res: CustomResponse<UserFlags>,
+    res: CustomResponse<
+      Pick<User, 'email' | 'firstName' | 'lastName' | 'address'> & UserFlags
+    >,
     next: NextFunction
   ) => {
     try {
+      if (!req.session.id || !req.session.user) {
+        req.session.destroy()
+        throw new Unauthorized('Unauthorized')
+      }
+
       const user = await this.deps.userService.getById(req.session.user.id)
 
       if (!user) {
@@ -38,6 +46,10 @@ export class UserController implements IUserController {
         success: true,
         message: 'User retrieved successfully',
         data: {
+          email: user.email,
+          firstName: user.firstName,
+          lastName: user.lastName,
+          address: user.address,
           needsWallet: !user.rapydWalletId,
           needsIDProof: !user.kycId
         }
